Add files via upload
You are creating a MotorVehicle superclass, and one subclass named MonsterTruck. You will then create two MonsterTruck objects, and display the information contained within those objects. You will also calculate the winning percentage for each truck, and display that as well. You will be demonstrating your knowledge of inheritance in Java with this assignment. 
MotorVehicle (superclass)
This class will have the following private data members:
•	truckYearBuilt, type int (i.e., 2010)
•	truckName, type String (i.e., Big Foot)
•	truckBodyType, type String (i.e., Blue)
•	truckPowerPlant, type String (i.e., Diesel)
It will also have a default constructor, and a constructor that takes all of the private data members as arguments. Create getters and setters for all data members, and also create a toString() method that prints out the MotorVehicle object details.
MonsterTruck (subclass)
This class will have the following private data members:
•	numberOfWins, type int
•	numberOfLosses, type int
•	specialManeuver, type String
It will also have a default constructor, and a constructor that takes all the private data members as arguments, to include a call to the superclass constructor. Create getters and setters for all data members, and also create a toString() method that prints out the MonsterTruck object, along with all of the information contained within MotorVehicle. You also need to perform a calculation to find the truck’s winning percentage. 
MonsterTruckRunner (class that contains main)
Create two monster truck objects and populate the data members. 
For the first monster truck, use the default constructor, followed by using setters to populate the values. Once set, use the getters to display the information, and include a call to the winning percentage calculation method.
For the second monster truck, use the constructor with all arguments to declare and initialize the object. Use the toString() method to display the result (hint: you will need to modify the toString() within MonsterTruck to also display the items from MotorVehicle). 
No user input is required for this assignment. 
Format the win percentage to two decimal places.